requirements.txt
fastapi
uvicorn
python-dotenv
neo4j
pydantic
requests
